STYLE: Remove `static_cast<unsigned int>` from dimension constants
Following ITK pull request https://github.com/InsightSoftwareConsortium/ITK/pull/5320 commit https://github.com/InsightSoftwareConsortium/ITK/commit/9e533c4bb5ac99a7edb72b2f7e39e8c00d69564d

+++ Common/OpenCL/Filters/itkGPUShrinkImageFilter.hxx
@@ -157,7 +157,7 @@ GPUShrinkImageFilter<TInputImage, TOutputImage>::GPUGenerateData()
     shrinkfactors[i] = factorSize[i];
   }
 
-  const auto ImageDim = static_cast<unsigned int>(InputImageDimension);
+  const auto ImageDim = InputImageDimension;
   switch (ImageDim)
   {
     case 1:
